Rename to BridgeAdmin, don't deploy BridgePool in whitelist token
